# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  request.RequestCreateContribution:
    properties:
      amount:
        type: number
    type: object
  request.RequestCreateGoal:
    properties:
      goal_amount:
        type: number
      name:
        type: string
      target_date:
        type: string
    type: object
  request.RequestUpdateContribution:
    properties:
      amount:
        type: number
    type: object
  request.RequestUpdateGoal:
    properties:
      goal_amount:
        type: number
      name:
        type: string
      target_date:
        type: string
    type: object
  response.ResponesContribution:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  response.ResponesGoal:
    properties:
      ads_by_amount:
        type: integer
      catalog_url:
        type: string
      goal_amount:
        type: number
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      target_date:
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  utils.JSONResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API Service of simple app for tracking your widescale goals
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoalTracker API
  version: "0.1"
paths:
  /v1/goal:
    get:
      description: Get goal by :id
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              allOf:
              - $ref: '#/definitions/utils.JSONResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/response.ResponesGoal'
                    type: array
                type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Goals Item
      tags:
      - Goals
    post:
      consumes:
      - application/json
      description: Create goal
      parameters:
      - description: Goal object for create
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/request.RequestCreateGoal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/response.ResponesGoal'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Goal Create
      tags:
      - Goals
  /v1/goal/:goal/contribution:
    get:
      description: Get contributions list
      parameters:
      - description: Goal ID for contribution
        in: path
        name: goal
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              allOf:
              - $ref: '#/definitions/utils.JSONResult'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/response.ResponesContribution'
                    type: array
                type: object
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Contribution List
      tags:
      - Contribution
    post:
      description: Create contribuitions
      parameters:
      - description: Goal ID for contribution
        in: path
        name: goal
        required: true
        type: integer
      - description: Contribution object for create
        in: body
        name: contribuition
        required: true
        schema:
          $ref: '#/definitions/request.RequestCreateContribution'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/response.ResponesContribution'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Contribution Create
      tags:
      - Contribution
  /v1/goal/:goal/contribution/:id:
    delete:
      description: Delete contribution by id
      parameters:
      - description: Goal ID for contribution
        in: path
        name: goal
        required: true
        type: integer
      - description: Contribution ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Contribution Delete
      tags:
      - Contribution
    put:
      description: Update contributions by id
      parameters:
      - description: Goal ID for contribution
        in: path
        name: goal
        required: true
        type: integer
      - description: Contribution ID
        in: path
        name: id
        required: true
        type: integer
      - description: Contribution`s fields for update
        in: body
        name: contribution
        required: true
        schema:
          $ref: '#/definitions/request.RequestUpdateContribution'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/response.ResponesContribution'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Contribution Update
      tags:
      - Contribution
  /v1/goal/:id:
    delete:
      description: Delete goal by id
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Goal Delete
      tags:
      - Goals
    get:
      description: Get goals list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/response.ResponesGoal'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Goal List
      tags:
      - Goals
    put:
      consumes:
      - application/json
      description: Update goal by id
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      - description: Goal`s fields for update
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/request.RequestUpdateGoal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/response.ResponesGoal'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Goal Update
      tags:
      - Goals
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"